Move validation in inputs.py
